requirements.txt
alembic==1.0.0
blinker==1.4
click==6.7
dominate==2.3.1
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-JSGlue==0.3.1
Flask-Migrate==2.2.1
Flask-Navigation==0.2.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
libusb1==1.6.5
M2Crypto==0.30.1
Mako==1.0.7
MarkupSafe==1.0
python-dateutil==2.7.3
python-dotenv==0.8.2
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.10
typing==3.6.4
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
